feat: perspective card bookmark changes (#1084)
* feat: perspective card bookmark changes

* feat: bookmark article paths without locale

* chore: use getPathDetails func to get language code as per feedback

* chore: feedback changes

* chore: feedback changes

---------
